Module Uutf.Manual

Manual sources and destinations.

Warning. Use only with `Manual decoder and encoders.

val src : decoder -> Bytes.t -> int -> int -> unit

src d s j l provides d with l bytes to read, starting at j in s. This byte range is read by calls to decode with d until `Await is returned. To signal the end of input call the function with l = 0.

val dst : encoder -> Bytes.t -> int -> int -> unit

dst e s j l provides e with l bytes to write, starting at j in s. This byte range is written by calls to encode with e until `Partial is returned. Use dst_rem to know the remaining number of non-written free bytes in s.

val dst_rem : encoder -> int

dst_rem e is the remaining number of non-written, free bytes in the last buffer provided with Manual.dst.